backend/requirements.txt
anyio==4.11.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
jwcrypto==1.5.6
mysqlclient==2.2.7
oauth2client==4.1.3
oauthlib==3.3.1
pillow==11.3.0
protobuf==6.33.0rc1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.2.5
Pyrebase4==4.8.0
python-jwt==4.1.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
